refactor: 불필요한 VO 제거
- 합당한 이유 없이 VO를 적용하였기에 분리하였음
